Inital extraction of utility functions into utils file
